+++ package.json
@@ -13,6 +13,7 @@
     "typegen": "sanity schema extract && sanity typegen generate"
   },
   "dependencies": {
+    "@emailjs/browser": "^4.4.1",
     "@sanity/assist": "^3.0.8",
     "@sanity/icons": "^3.4.0",
     "@sanity/image-url": "^1.0.2",
